# add ZLIB/QUICKLZ/LZMA compress ratio and performance to test