The compress-test outputs as:
```
TOKU_ZLIB_METHOD Time=0.963718s , Ratio=0.34[8947327/26667027]
TOKU_ZLIB_WITHOUT_CHECKSUM_METHOD Time=0.947162s, Ratio=0.34[8935295/26667027]
TOKU_QUICKLZ_METHOD Time=1.734388s, Ratio=0.35[9304195/26667027]
TOKU_LZMA_METHOD Time=5.949371s, Ratio=0.34[9089546/26667027]
```

So, we can easy to measure the performance for different compression methods from the test.